+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Purchase improvements
     * Grouped specializations in selection popup by in and out of career
     * Grouped force powers in selection popup  by required force rating
+    * Spending XP now whispers the GM to notify them (this can be disabled in the system settings)
 * Fixes:
   * Removed the "purchase" context menu option when editing a Specialization that is not attached to a character
   * Improves nested tag formatting (e.g., allows for strings like `[B]Average ([di][di]) Skill[b]` to properly render)